Constant-time stepping not so constant.
Good advantage on a single motion with thousands of phrases.
Barely better on motions with ~10 phrases (assuming we use double for Time in both cases).
Requires double for Time type to be accurate.
Slower when doing many motions with 2 phrases.
Probably the forward-only memory traversal of the naive Sequence approach is cache-friendlier.

+++ src/choreograph/Source.hpp
@@ -35,7 +35,7 @@ namespace choreograph
 /// Define our Time type here so it's easier to change out if needed.
 /// Float loses precision pretty quickly, but is fast and doesn't take up much space.
 /// Double works nicely when adding together lots of little bits.
-using Time = double;
+using Time = float;
 
 template<typename T>
 class Sequence;
